removed redundant version
