--- component/models/sdjwt/common/v5.go
@@ -1,84 +1,0 @@
-/*
-Copyright Avast Software. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package common
-
-import (
-	"fmt"
-	afgjwt "github.com/hyperledger/aries-framework-go/component/models/jwt"
-	utils "github.com/hyperledger/aries-framework-go/component/models/util/maphelpers"
-)
-
-type commonV5 struct {
-}
-
-func newCommonV5() *commonV5 {
-	return &commonV5{}
-}
-
-// VerifyDisclosuresInSDJWT checks for disclosure inclusion in SD-JWT.
-func (c *commonV5) VerifyDisclosuresInSDJWT(disclosures []string, signedJWT *afgjwt.JSONWebToken) error {
-	claims := utils.CopyMap(signedJWT.Payload)
-
-	// check that the _sd_alg claim is present
-	// check that _sd_alg value is understood and the hash algorithm is deemed secure.
-	cryptoHash, err := GetCryptoHashFromClaims(claims)
-	if err != nil {
-		return err
-	}
-
-	var disclosuresClaims []*DisclosureClaim
-
-	for _, disclosure := range disclosures {
-		digest, err := GetHash(cryptoHash, disclosure)
-		if err != nil {
-			return err
-		}
-
-		found, err := isDigestInClaims(digest, claims)
-		if err != nil {
-			return err
-		}
-
-		if found {
-			continue
-		}
-
-		if disclosuresClaims == nil {
-			disclosuresClaims, err = GetDisclosureClaims(disclosures)
-			if err != nil {
-				return fmt.Errorf("getDisclosureClaims: %w", err)
-			}
-		}
-
-		// Check if given digest contains in nested disclosures
-		if found = isDigestInDisclosures(disclosuresClaims, digest); found {
-			continue
-		}
-
-		return fmt.Errorf("disclosure digest '%s' not found in SD-JWT disclosure digests", digest)
-	}
-
-	return nil
-}
-
-func isDigestInDisclosures(disclosuresClaims []*DisclosureClaim, digest string) bool {
-	for _, parsedDisclosure := range disclosuresClaims {
-		if parsedDisclosure.Type != DisclosureClaimTypeObject {
-			continue
-		}
-		found, err := isDigestInClaims(digest, parsedDisclosure.Value.(map[string]interface{}))
-		if err != nil {
-			return false
-		}
-
-		if found {
-			return true
-		}
-	}
-
-	return false
-}
